requirements.txt
openai
python-dotenv
colorama
graphviz
requests
jupyter
